+++ src/cli.rs
@@ -93,8 +93,8 @@ pub enum Commands {
     #[command(about = "Get the path of a project")]
     Path(PathCommand),
     // Future:
-    // #[command(aliases = &["ter", "shell", "term", "sh", "cmd"], about = "Open a project in the terminal")]
-    // Terminal(TerminalCommand),
+    #[command(aliases = &["ter", "shell", "term", "sh", "cmd"], about = "Open a project in the terminal")]
+    Terminal(TerminalCommand),
     // #[command(aliases = &["structure"], about = "Show the project structure")]
     // Tree(TreeCommand),
     // #[command(aliases = &["label"], about = "Manage tags for a project")]
