+++ src/commands/terminal.rs
@@ -1,0 +1,13 @@
+
+#[derive(clap::Args)]
+pub struct TerminalCommand {
+    pub name: String,
+}
+
+impl TerminalCommand {
+    pub fn handle(self) -> Result<(), Box<dyn std::error::Error>> {
+       
+
+        Ok(())
+    }
+}
